# contracts/openapi/vvoice.openapi.json
{
  "components": {
    "schemas": {
      "AsrJobCleanupResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          },
          "job_ids": {
            "items": {
              "type": "string"
            },
            "title": "Job Ids",
            "type": "array"
          }
        },
        "required": [
          "deleted",
          "job_ids"
        ],
        "title": "AsrJobCleanupResponse",
        "type": "object"
      },
      "AsrJobDeleteResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "job_id"
        ],
        "title": "AsrJobDeleteResponse",
        "type": "object"
      },
      "AsrJobResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "cancellation_mode": {
            "const": "safe_point",
            "title": "Cancellation Mode",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Reason"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "progress_stage": {
            "enum": [
              "queued",
              "preparing_input",
              "running_model",
              "finalizing",
              "retry_wait",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Progress Stage",
            "type": "string"
          },
          "raw_segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegmentResponse"
            },
            "title": "Raw Segments",
            "type": "array"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "sample_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegmentResponse"
            },
            "title": "Segments",
            "type": "array"
          },
          "stage_started_at": {
            "title": "Stage Started At",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "timing_status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Timing Status",
            "type": "string"
          },
          "transcript_edited": {
            "title": "Transcript Edited",
            "type": "boolean"
          },
          "transcript_revision": {
            "title": "Transcript Revision",
            "type": "integer"
          },
          "transcript_updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Updated At"
          }
        },
        "required": [
          "job_id",
          "status",
          "filename",
          "language",
          "created_at",
          "started_at",
          "completed_at",
          "error",
          "attempt",
          "max_attempts",
          "cancel_requested",
          "cancellation_mode",
          "failed_reason",
          "progress_stage",
          "stage_started_at",
          "text",
          "sample_rate",
          "duration_seconds",
          "audio_url",
          "raw_text",
          "raw_segments",
          "segments",
          "timing_status",
          "transcript_revision",
          "transcript_edited",
          "transcript_updated_at"
        ],
        "title": "AsrJobResponse",
        "type": "object"
      },
...
      "Body_create_asr_job_api_v1_asr_jobs_post": {
        "properties": {
          "audio": {
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "type": "string"
          },
          "language": {
            "default": "vi",
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "audio"
        ],
        "title": "Body_create_asr_job_api_v1_asr_jobs_post",
        "type": "object"
      },
...
      "Body_transcribe_audio_api_v1_asr_transcribe_post": {
        "properties": {
          "audio": {
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "type": "string"
          },
          "language": {
            "default": "vi",
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "audio"
        ],
        "title": "Body_transcribe_audio_api_v1_asr_transcribe_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TranscriptRevisionRequest": {
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TranscriptSegmentEdit"
                },
                "maxItems": 5000,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 100000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "TranscriptRevisionRequest",
        "type": "object"
      },
      "TranscriptSegmentEdit": {
        "properties": {
          "segment_id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Segment Id",
            "type": "string"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "segment_id",
          "text"
        ],
        "title": "TranscriptSegmentEdit",
        "type": "object"
      },
      "TranscriptSegmentResponse": {
        "properties": {
          "end_seconds": {
            "title": "End Seconds",
            "type": "number"
          },
          "segment_id": {
            "title": "Segment Id",
            "type": "string"
          },
          "start_seconds": {
            "title": "Start Seconds",
            "type": "number"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "segment_id",
          "start_seconds",
          "end_seconds",
          "text"
        ],
        "title": "TranscriptSegmentResponse",
        "type": "object"
      },
      "TranscriptionResponse": {
        "properties": {
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "sample_rate": {
            "title": "Sample Rate",
            "type": "integer"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegmentResponse"
            },
            "title": "Segments",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "timing_status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Timing Status",
            "type": "string"
          }
        },
        "required": [
          "text",
          "sample_rate",
          "duration_seconds",
          "timing_status",
          "segments"
        ],
        "title": "TranscriptionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VassilStudio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/asr/jobs": {
      "delete": {
        "operationId": "cleanup_asr_jobs_api_v1_asr_jobs_delete",
        "parameters": [
          {
            "in": "query",
            "name": "max_age_seconds",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrJobCleanupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup Asr Jobs",
        "tags": [
          "asr"
        ]
      },
      "get": {
        "operationId": "list_asr_jobs_api_v1_asr_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AsrJobResponse"
                  },
                  "title": "Response List Asr Jobs Api V1 Asr Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Asr Jobs",
        "tags": [
          "asr"
        ]
      },
      "post": {
        "operationId": "create_asr_job_api_v1_asr_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_asr_job_api_v1_asr_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Idempotency key conflicts with an earlier request"
          },
          "413": {
            "description": "Decoded audio exceeds processing limits"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Job queue is full"
          }
        },
        "summary": "Create Asr Job",
        "tags": [
          "asr"
        ]
      }
    },
    "/api/v1/asr/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_asr_job_api_v1_asr_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrJobDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Asr Job",
        "tags": [
          "asr"
        ]
      },
      "get": {
        "operationId": "get_asr_job_api_v1_asr_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asr Job",
        "tags": [
          "asr"
        ]
      }
    },
    "/api/v1/asr/jobs/{job_id}/audio": {
      "get": {
        "operationId": "get_asr_job_audio_api_v1_asr_jobs__job_id__audio_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asr Job Audio",
        "tags": [
          "asr"
        ]
      }
    },
    "/api/v1/asr/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_asr_job_api_v1_asr_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Asr Job",
        "tags": [
          "asr"
        ]
      }
    },
    "/api/v1/asr/jobs/{job_id}/exports/{export_format}": {
      "get": {
        "operationId": "export_asr_transcript_api_v1_asr_jobs__job_id__exports__export_format__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "export_format",
            "required": true,
            "schema": {
              "enum": [
                "txt",
                "srt",
                "vtt",
                "json"
              ],
              "title": "Export Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Transcript or timed segments are not ready"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Asr Transcript",
        "tags": [
          "asr"
        ]
      }
    },
    "/api/v1/asr/jobs/{job_id}/transcript": {
      "patch": {
        "operationId": "revise_asr_transcript_api_v1_asr_jobs__job_id__transcript_patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Transcript is not ready or the revision is stale"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revise Asr Transcript",
        "tags": [
          "asr"
        ]
      }
    },
    "/api/v1/asr/transcribe": {
      "post": {
        "operationId": "transcribe_audio_api_v1_asr_transcribe_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_api_v1_asr_transcribe_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcribe Audio",
        "tags": [
          "asr"
        ]
      }
    },
...
  }
}